pyproject.toml
[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "dlite_entities_service"

[project]
name = "dlite-entities-service"
authors = [
    {name = "Casper Welzel Andersen", email = "casper.w.andersen@sintef.no"},
]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Natural Language :: English",
    "Operating System :: OS Independent",
]
keywords = ["dlite", "mongodb"]
requires-python = "~=3.10"
dynamic = ["version", "description"]

dependencies = [
    "fastapi ~=0.108.0",
    "pydantic-settings ~=2.1",
    "pymongo ~=4.6",
    "python-dotenv ~=1.0",
    "uvicorn >=0.24.0,<1",
]

[project.optional-dependencies]
cli = [
    "typer[all] >=0.9.0,<1",
]
testing = [
    "dlite-python ~=0.4.5; python_version < '3.12'",
    "httpx >=0.25.1,<1",
    "mongomock ~=4.1",
    "pytest ~=7.4",
    "pytest-cov ~=4.1",
    "pyyaml ~=6.0",
    "dlite-entities-service[cli]",
]
dev = [
    "pre-commit ~=3.6",
    "dlite-entities-service[testing,cli]",
]

[project.scripts]
entities-service = "dlite_entities_service.cli.main:APP"

[project.urls]
Home = "https://github.com/CasperWA/dlite-entities-service"
Documentation = "https://CasperWA.github.io/dlite-entities-service"
Source = "https://github.com/CasperWA/dlite-entities-service"
"Issue Tracker" = "https://github.com/CasperWA/dlite-entities-service/issues"
Changelog = "https://CasperWA.github.io/dlite-entities-service/latest/CHANGELOG"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
scripts_are_modules = true
warn_unused_configs = true
show_error_codes = true
allow_redefinition = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.ruff.lint]
extend-select = [
  "E",  # pycodestyle
  "F",  # pyflakes
  "B",  # flake8-bugbear
  "BLE",  # flake8-blind-except
  "I",  # isort
  "ARG",  # flake8-unused-arguments
  "C4",  # flake8-comprehensions
  "ICN",  # flake8-import-conventions
  "G",  # flake8-logging-format
  "PGH",  # pygrep-hooks
  "PIE",  # flake8-pie
  "PL",  # pylint
  "PT",  # flake8-pytest-style
  "PTH",  # flake8-use-pathlib
  "RET",  # flake8-return
  "RUF",  # Ruff-specific
  "SIM",  # flake8-simplify
  "T20",  # flake8-print
  "YTT",  # flake8-2020
  "EXE",  # flake8-executable
  "PYI",  # flake8-pyi
]
ignore = [
  "PLR",  # Design related pylint codes
  "B008",  # Performing function calls in argument defaults - done all the time in the CLI.
]
isort.required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "BLE",  # flake8-blind-except
    "T20",  # flake8-print
]

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-rs --cov=dlite_entities_service --cov-report=term-missing:skip-covered --no-cov-on-fail"
filterwarnings = [
    # Treat all warnings as errors
    "error",

    # mongomock uses pkg_resources
    "ignore:.*pkg_resources is deprecated as an API.*:DeprecationWarning",
]
